refactoring and cleaning up
