+++ .eslintrc
@@ -8,7 +8,8 @@
     no-param-reassign: [
       "error",
       { "props": false }
-    ]
+    ],
+    "no-console": 0
   },
   "globals": {
     "describe": true,
