+++ package.json
@@ -5,10 +5,9 @@
   "main": "./dist/index.js",
   "scripts": {
     "test": "gulp test",
-    "cov": "istanbul cover _mocha -- test/",
     "test-watch": "gulp watch",
-    "coveralls": "istanbul cover _mocha && istanbul-coveralls",
-    "docs": "docco src/*.js && node script/generate_feature_doc",
+    "coveralls": "gulp coveralls",
+    "docs": "gulp docs",
     "build": "gulp build"
   },
   "directories": {
@@ -25,7 +24,6 @@
     "babel-preset-es2015": "^6.9.0",
     "chai": "^3.5.0",
     "dirty-chai": "^1.2.2",
-    "docco": "^0.7.0",
     "eslint": "^3.1.1",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.11.1",
@@ -34,14 +32,14 @@
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
     "gulp-concat": "^2.6.0",
+    "gulp-coveralls": "^0.1.4",
+    "gulp-docco": "0.0.4",
     "gulp-eslint": "^3.0.1",
     "gulp-exclude-gitignore": "^1.0.0",
     "gulp-istanbul": "^1.0.0",
     "gulp-mocha": "^2.2.0",
     "gulp-nsp": "^2.4.2",
     "gulp-plumber": "^1.1.0",
-    "istanbul": "0.4.3",
-    "istanbul-coveralls": "1.0.3",
     "mocha": "2.4.5",
     "sinon": "1.17.3",
     "sinon-chai": "^2.8.0"
